Withdraw handler implemented

+++ src/main/java/com/boris/fundingarbitrage/model/assetops/Withdrawal.java
@@ -4,7 +4,7 @@ import com.boris.fundingarbitrage.model.Validations;
 import com.boris.fundingarbitrage.model.exchange.WalletAddress;
 import lombok.NonNull;
 
-public record Withdrawal(int amount, double fee, @NonNull WalletAddress address) {
+public record Withdrawal(double amount, double fee, @NonNull WalletAddress address) {
 	public Withdrawal {
 		Validations.requirePositive(amount, "Amount");
 	}
